Moved all but Substring into the StringValue class and it's working well. Added the asValue() method to Column to make it easier to get an appropriate DBValue type.
No longer sure that *Variable is a good idea, maybe I can get StringColumn to subclass StringValue and it will all be easy ... maybe ...
Also DBMath is now looking kind of awkward and there are lots of inner classes in StringValue that might be better else where.
Substring is a pain by the way.