NumberExpression now does maths with a beautiful stream like interface.
The *Variable interfaces are now *Return and that seems to reflect their function better.  They also appear to be here to stay.

The plan is to  move the operators into BooleanExpression so that addComparison will take a single BooleanExpression

Still not sure about all those DB*naryFunction inner classes but I really think they need to be generalised made safe and locked down in their own *.internal.functions package